requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
schedule==1.2.1
psycopg2-binary==2.9.9
openai==1.35.0
pdfplumber==0.11.0
